continue to add CPI cases and minor fix some codes

+++ remote-scripts/CPI-NETWORK-RESOURCE-GROUP-NAME.py
@@ -25,6 +25,9 @@ def Update_cf_manifest(source_yml_file, destination_yml_file, publicIP):
     out = load_manifest(source_yml_file)
     out['jobs'] = remove_unnessary_jobs(out)
     out['releases'] = remove_unnessary_releases(out)
+    out['resource_pools'] = remove_unnessary_resourcepools(out)
+    out['releases'][0]['version'] = 'latest'
+    out['resource_pools'][0]['stemcell']['version'] = 'latest'
     network = filter(lambda j:j.get('name') == 'cf_public', out['networks'])
     network[0]['cloud_properties'] = {'resource_group_name':temp_resource_group_name}
     out['name'] = cf_deployment_name
@@ -44,19 +47,24 @@ def RunTest():
     publicIP = create_publicIP_resource(temp_resource_group_name)
     Update_cf_manifest(source_manifest_name, destination_manifest_name, publicIP)
     if DeployCF(destination_manifest_name):
-        ResultLog.info('PASS')
-        RunLog.info("Remove deployed CF %s" % cf_deployment_name)		
-        Run("echo yes | bosh delete deployment %s" % cf_deployment_name)
-        UpdateState("TestCompleted")
+        RunLog.info('Start to verify the publicIP resource')
+        host_name = Run("bosh vms --details | grep separatenetwork_z1 | awk '{print $13}'")
+        publicIP_info = Run('az network public-ip show -g %s -n publicIP --query ipConfiguration.id' % temp_resource_group_name)
+        if host_name and (host_name.strip('\n') in publicIP_info):
+            RunLog.info("Test PASS, Remove deployed CF %s and test resources" % cf_deployment_name)		
+            Run("echo yes | bosh delete deployment %s" % cf_deployment_name)
+            Run('az group delete --name %s -y' % temp_resource_group_name)
+            ResultLog.info('PASS')
+        else:
+            RunLog.info("Test FAIL")
+            ResultLog.info('FAIL')
     else:
         ResultLog.error('FAIL')
-        UpdateState("TestCompleted")
-    Run('az group delete --name %s -y' % temp_resource_group_name)
-
+    UpdateState("TestCompleted")
 
 #set variable
 source_manifest_name = 'example_manifests/single-vm-cf.yml'
-destination_manifest_name = 'cpi-vip-rg-single-vm-cf.yml'
+destination_manifest_name = 'cpi-separate-network-cf.yml'
 cf_deployment_name = 'separate-network-cf'
 temp_resource_group_name = 'testrg-%s' % random.randint(1,100)
